% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{choose_window_size}
\alias{choose_window_size}
\title{Choose a rolling window size based on number of samples.}
\usage{
choose_window_size(n)
}
\description{
Choose a rolling window size based on number of samples.
}
\keyword{internal}
