% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcuo.R
\docType{package}
\name{dcuo-package}
\alias{dcuo}
\alias{dcuo-package}
\title{dcuo: Combat log parsing tools for DC Universe Online}
\description{
dcuo: Combat log parsing tools for DC Universe Online
}
\author{
\strong{Maintainer}: Greg Lin \email{gglliinn@gmail.com}

}
