% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcuo.R
\docType{package}
\name{dcuolog-package}
\alias{dcuolog}
\alias{dcuolog-package}
\title{dcuolog: Combat log parsing tools for DC Universe Online}
\description{
dcuolog: Combat log parsing tools for DC Universe Online
}
\author{
\strong{Maintainer}: Greg Lin \email{glin@glin.io}

}
