% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{fix_attribution}
\alias{fix_attribution}
\title{Fix incorrect attribution of abilities with aura or chain reaction effects.}
\usage{
fix_attribution(combat_events)
}
\description{
Fix incorrect attribution of abilities with aura or chain reaction effects.
}
\keyword{internal}
