% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_numeric_groups}
\alias{get_numeric_groups}
\title{Get names of numeric capture groups from a pattern.}
\usage{
get_numeric_groups(pattern)
}
\description{
Get names of numeric capture groups from a pattern.
}
\keyword{internal}
