% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_combat}
\alias{parse_combat}
\title{Parse general combat events: damage, heal, power, absorb, supercharge.}
\usage{
parse_combat(log, file = NULL, time_digits = 3)
}
\arguments{
\item{log}{Character vector of log lines.}

\item{file}{Name of a log file to read from.}

\item{time_digits}{Number of decimal places to use for time (seconds).}
}
\value{
A \code{combat_events} object.
}
\description{
Parse general combat events: damage, heal, power, absorb, supercharge.
}
