% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_dcuo}
\alias{parse_dcuo}
\title{Parse a DCUO combat log.}
\usage{
parse_dcuo(log, file = NULL, time_digits = 3, type = c("combat",
  "summary", "crowd_control", "knockout", "dodge"))
}
\arguments{
\item{log}{Character vector of log lines.}

\item{file}{Name of a log file to read from.}

\item{time_digits}{Number of decimal places to use for time (seconds).}

\item{type}{One of "combat", "summary", "crowd_control", "knockout", "dodge" (combat by default).}
}
\value{
Table of parsed combat log entries.
}
\description{
Parse a DCUO combat log.
}
