% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_dodge}
\alias{parse_dodge}
\title{Parse dodges.}
\usage{
parse_dodge(log, file = NULL, time_digits = 3)
}
\arguments{
\item{log}{Character vector of log lines.}

\item{file}{Name of a log file to read from.}

\item{time_digits}{Number of decimal places to use for time (seconds).}
}
\value{
A \code{dodge_events} object.
}
\description{
Parse dodges.
}
