% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_log}
\alias{parse_log}
\title{Parse a combat log given a pattern.}
\usage{
parse_log(log, pattern, file = NULL, time_digits = 3, class = "events")
}
\description{
Parse a combat log given a pattern.
}
\keyword{internal}
