% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_summary}
\alias{parse_summary}
\title{Parse combat parser summary entries.}
\usage{
parse_summary(log, file = NULL, time_digits = 3)
}
\arguments{
\item{log}{Character vector of log lines.}

\item{file}{Name of a log file to read from.}

\item{time_digits}{Number of decimal places to use for time (seconds)}
}
\value{
A \code{parser_summary} object.
}
\description{
Parse combat parser summary entries.
}
