% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{plot.combat_ts}
\alias{plot.combat_ts}
\title{Plot combat time series data.}
\usage{
\method{plot}{combat_ts}(ts)
}
\arguments{
\item{ts}{A \code{combat_ts} object.}
}
\description{
Plot combat time series data.
}
