% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_lines}
\alias{read_lines}
\title{Read text lines from a file.}
\usage{
read_lines(file)
}
\arguments{
\item{file}{Name of the file to read.}
}
\value{
Character vector of text lines in UTF-8.
}
\description{
Read text lines from a file.
}
