% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_dt_order}
\alias{set_dt_order}
\title{Like \code{data.table::setcolorder}, but the column order can be a superset of
columns in the \code{data.table}.}
\usage{
set_dt_order(dt, neworder)
}
\description{
Like \code{data.table::setcolorder}, but the column order can be a superset of
columns in the \code{data.table}.
}
\keyword{internal}
