% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{smooth.combat_ts}
\alias{smooth.combat_ts}
\title{Smooth combat time series data using a Gaussian filter window.}
\usage{
\method{smooth}{combat_ts}(ts)
}
\arguments{
\item{ts}{A \code{combat_ts} object.}
}
\value{
The smoothed \code{combat_ts} object.
}
\description{
Smooth combat time series data using a Gaussian filter window.
}
