% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_dt}
\alias{sort_dt}
\title{Sort a data table in place.}
\usage{
sort_dt(dt, by, order = c("desc", "asc"))
}
\arguments{
\item{dt}{A data table.}

\item{sort_by}{Column to sort by.}

\item{sort_order}{Sort in "desc" or "asc" order (desc by default).}
}
\description{
Sort a data table in place.
}
\keyword{internal}
