% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_match_named}
\alias{str_match_named}
\title{Extract named capture groups from a string.}
\usage{
str_match_named(string, pattern)
}
\description{
Extract named capture groups from a string.
}
\keyword{internal}
