% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.combat_events}
\alias{summary.combat_events}
\title{Summarize general combat events.}
\usage{
\method{summary}{combat_events}(events, type = c("damage", "healing", "power",
  "absorb", "supercharge", "damage_in", "healing_in", "power_in"),
  by_ability = FALSE, power_metrics = FALSE, combat_window = 10)
}
\arguments{
\item{events}{A \code{combat_events} object.}

\item{type}{One of "damage", "healing", "power", "absorb", "supercharge",
"damage_in", "healing_in", "power_in" (damage by default).}

\item{by_ability}{Aggregate by ability.}

\item{power_metrics}{Include power in metrics (only for damage, healing, power, or absorb).}

\item{combat_window}{Maximum window of combat time (10 seconds by default).}
}
\value{
Summary table of general combat events.
}
\description{
Summarize general combat events.
}
