% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.crowd_control_events}
\alias{summary.crowd_control_events}
\title{Summarize crowd control events.}
\usage{
\method{summary}{crowd_control_events}(events, by = c("source", "target",
  "ability", "effect"))
}
\arguments{
\item{events}{A \code{crowd_control_events} object.}

\item{by}{Group by any combination of "source", "target", "ability", "effect" (all by default).}
}
\value{
Summary table of crowd control events.
}
\description{
Summarize crowd control events.
}
