% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.dodge_events}
\alias{summary.dodge_events}
\title{Summarize dodge events.}
\usage{
\method{summary}{dodge_events}(events, by = c("source", "target", "ability"))
}
\arguments{
\item{events}{A \code{dodge_events} object.}

\item{by}{Group by any combination of "source", "target", "ability" (all by default).}
}
\value{
Summary table of dodge events.
}
\description{
Summarize dodge events.
}
