% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.knockout_events}
\alias{summary.knockout_events}
\title{Summarize knockout events.}
\usage{
\method{summary}{knockout_events}(events, by = c("source", "target"))
}
\arguments{
\item{events}{A \code{knockout_events} object.}

\item{by}{Group by any combination of "source", "target" (all by default).}
}
\value{
Summary table of knockout events.
}
\description{
Summarize knockout events.
}
