% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.parser_summary}
\alias{summary.parser_summary}
\title{Summarize combat parser summary events.}
\usage{
\method{summary}{parser_summary}(events, type = c("damage", "healing",
  "power"))
}
\arguments{
\item{events}{A \code{parser_summary} object.}

\item{type}{One of "damage", "healing", "power" (damage by default).}
}
\value{
Summary table of combat summary events.
}
\description{
Summarize combat parser summary events.
}
