% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries.combat_events}
\alias{timeseries.combat_events}
\title{Create a time series (1 second interval) from combat events.}
\usage{
\method{timeseries}{combat_events}(events, type = c("damage", "healing",
  "power", "absorb", "supercharge", "damage_in", "healing_in", "power_in"),
  by_ability = FALSE)
}
\arguments{
\item{events}{A \code{combat_events} object.}

\item{type}{One of "damage", "healing", "power", "absorb", "supercharge",
"damage_in", "healing_in", "power_in" (damage by default).}

\item{by_ability}{Aggregate by ability.}
}
\value{
A \code{combat_ts} object.
}
\description{
Create a time series (1 second interval) from combat events.
}
