% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top.R
\name{top.combat_events}
\alias{top.combat_events}
\title{Show top events in the combat log.}
\usage{
\method{top}{combat_events}(events, n = 30, type = c("damage", "healing",
  "power", "absorb", "supercharge"))
}
\arguments{
\item{events}{A \code{combat_events} object.}

\item{n}{Number of events (30 by default).}

\item{type}{One of "damage", "healing", "power", "absorb", "supercharge" (damage by default).}
}
\value{
The top n events in the log.
}
\description{
Show top events in the combat log.
}
