% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{total_time}
\alias{total_time}
\title{Find the total elapsed time for a vector of date-time objects.}
\usage{
total_time(time, units = "secs")
}
\description{
Find the total elapsed time for a vector of date-time objects.
}
\keyword{internal}
