% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{wma_gaussian}
\alias{wma_gaussian}
\title{Weighted moving average using a Gaussian filter window.}
\usage{
wma_gaussian(x, window_size, sd = 0.5)
}
\description{
Weighted moving average using a Gaussian filter window.
}
\keyword{internal}
