% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.combat_events}
\alias{summary.combat_events}
\title{Summarize general combat events.}
\usage{
\method{summary}{combat_events}(events, type = c("damage", "healing", "power",
  "absorb", "supercharge"), by = "source", power_metrics = FALSE,
  active_window = 10)
}
\arguments{
\item{events}{A \code{combat_events} object.}

\item{type}{One of "damage", "healing", "power", "absorb", "supercharge" (damage by default).}

\item{by}{Group by any combination of "source", "target", "ability" (source by default).}

\item{power_metrics}{Include power in metrics (only valid when grouping by source).}

\item{active_window}{Window of active combat time (10 seconds by default).}
}
\value{
Summary table of general combat events.
}
\description{
Summarize general combat events.
}
